+++ forui/test/src/widgets/badge_golden_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:forui/src/widgets/badge/badge.dart';
+import 'package:forui/src/widgets/badge.dart';
 import '../test_scaffold.dart';
 
 void main() {
@@ -26,7 +26,7 @@ void main() {
       ('destructive', FBadgeVariant.destructive),
       ('outline', FBadgeVariant.outline),
     ]) {
-      testWidgets('${theme.name} with FBadgeContent', (tester) async {
+      testWidgets('${theme.name} with content', (tester) async {
         await tester.pumpWidget(
           TestScaffold(
             theme: theme.data,