+++ forui/test/src/widgets/tabs/tabs_golden_test.dart
@@ -33,22 +33,8 @@ void main() {
           theme: theme.data,
           child: FTabs(
             children: [
-              FTabEntry(
-                label: const Text('Account'),
-                child: FCard(
-                  title: const Text('Account'),
-                  subtitle: const Text('Make changes to your account here. Click save when you are done.'),
-                  child: Container(color: Colors.blue, height: 100),
-                ),
-              ),
-              FTabEntry(
-                label: const Text('Password'),
-                child: FCard(
-                  title: const Text('Password'),
-                  subtitle: const Text('Change your password here. After saving, you will be logged out.'),
-                  child: Container(color: Colors.red, height: 100),
-                ),
-              ),
+              FTabEntry(label: const Text('Account'), child: Container(color: Colors.blue, height: 100)),
+              FTabEntry(label: const Text('Password'), child: Container(color: Colors.red, height: 100)),
             ],
           ),
         ),
@@ -102,22 +88,8 @@ void main() {
           theme: theme.data,
           child: FTabs(
             children: [
-              FTabEntry(
-                label: const Text('Account'),
-                child: FCard(
-                  title: const Text('Account'),
-                  subtitle: const Text('Make changes to your account here. Click save when you are done.'),
-                  child: Container(color: Colors.blue, height: 100),
-                ),
-              ),
-              FTabEntry(
-                label: const Text('Password'),
-                child: FCard(
-                  title: const Text('Password'),
-                  subtitle: const Text('Change your password here. After saving, you will be logged out.'),
-                  child: Container(color: Colors.red, height: 100),
-                ),
-              ),
+              FTabEntry(label: const Text('Account'), child: Container(color: Colors.blue, height: 100)),
+              FTabEntry(label: const Text('Password'), child: Container(color: Colors.red, height: 100)),
             ],
           ),
         ),
@@ -140,20 +112,8 @@ void main() {
             child: FTabs(
               expands: true,
               children: [
-                FTabEntry(
-                  label: const Text('Account'),
-                  child: FCard(
-                    title: const Text('Account'),
-                    subtitle: const Text('Make changes to your account here. Click save when you are done.'),
-                  ),
-                ),
-                FTabEntry(
-                  label: const Text('Password'),
-                  child: FCard(
-                    title: const Text('Password'),
-                    subtitle: const Text('Change your password here. After saving, you will be logged out.'),
-                  ),
-                ),
+                FTabEntry(label: const Text('Account'), child: Container(color: Colors.blue, height: 100)),
+                FTabEntry(label: const Text('Password'), child: Container(color: Colors.red, height: 100)),
               ],
             ),
           ),