+++ spotlight/lib/widgets/sidebar.dart
@@ -39,7 +39,7 @@ class Sidebar extends StatelessWidget {
             ),
             footer: Padding(
               padding: const .symmetric(horizontal: 16),
-              child: FCard.raw(
+              child: FCard(
                 child: Padding(
                   padding: const .symmetric(vertical: 12, horizontal: 16),
                   child: Row(